# replace histogram with base-2 implementation