--- a/heatmap/src/error.rs
+++ b/heatmap/src/error.rs
@@ -2,11 +2,11 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
-use rustcommon_histogram::HistogramError;
+use histogram::Error as HistogramError;
 use thiserror::Error;
 
-#[derive(Error, Debug, PartialEq)]
-pub enum HeatmapError {
+#[derive(Error, Debug, PartialEq, Eq)]
+pub enum Error {
     #[error("heatmap contains no samples")]
     Empty,
     #[error("invalid percentile")]
@@ -15,7 +15,7 @@ pub enum HeatmapError {
     OutOfRange,
 }
 
-impl From<HistogramError> for HeatmapError {
+impl From<HistogramError> for Error {
     fn from(other: HistogramError) -> Self {
         match other {
             HistogramError::Empty => Self::Empty,