--- a/heatmap/src/heatmap.rs
+++ b/heatmap/src/heatmap.rs
@@ -2,14 +2,11 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
-use crate::Heatmap;
-use crate::HeatmapError;
-use core::sync::atomic::AtomicU64;
-use core::sync::atomic::AtomicUsize;
+use crate::Error;
+use crate::*;
+use core::sync::atomic::*;
 
-use rustcommon_atomics::{Atomic, Ordering};
-use rustcommon_histogram::{AtomicCounter, AtomicHistogram, Counter, Indexing};
-use rustcommon_time::*;
+use histogram::{Bucket, Histogram};
 
 /// AtomicHeatmaps are concurrent datastructures which store counts for
 /// timestamped values over a configured time range with individual histograms
@@ -17,21 +14,15 @@ use rustcommon_time::*;
 /// buffer, unless they are newer than that slice may hold. When this happens,
 /// old slices are cleared and reused. This configuration results in a fully
 /// pre-allocated datastructure with concurrent read-write access.
-pub struct AtomicHeatmap<Value, Count> {
-    slices: Vec<AtomicHistogram<Value, Count>>,
+pub struct Heatmap {
+    slices: Vec<Histogram>,
     current: AtomicUsize,
-    next_tick: Instant<Nanoseconds<AtomicU64>>,
-    resolution: Duration<Nanoseconds<u64>>,
-    summary: AtomicHistogram<Value, Count>,
+    next_tick: AtomicInstant,
+    resolution: Duration,
+    summary: Histogram,
 }
 
-impl<Value, Count> AtomicHeatmap<Value, Count>
-where
-    Value: Indexing,
-    Count: AtomicCounter + Default,
-    u64: From<Value> + From<<Count as Atomic>::Primitive>,
-    <Count as Atomic>::Primitive: Copy,
-{
+impl Heatmap {
     /// Create a new `AtomicHeatmap` which can store values up and including the
     /// `max` while maintaining precision across a wide range of values. The
     /// `precision` is expressed in the number of significant figures preserved.
@@ -42,27 +33,22 @@ where
     /// be slightly longer than the requested span. Smaller durations for the
     /// resolution cause more memory to be used, but a smaller batches of
     /// samples to age out at each time step.
-    pub fn new(
-        max: Value,
-        precision: u8,
-        span: Duration<Nanoseconds<u64>>,
-        resolution: Duration<Nanoseconds<u64>>,
-    ) -> Self {
+    pub fn new(m: u32, r: u32, n: u32, span: Duration, resolution: Duration) -> Self {
         let mut slices = Vec::new();
-        let mut true_span = Duration::<Nanoseconds<u64>>::from_nanos(0);
+        let mut true_span = Duration::from_nanos(0);
         while true_span < span {
-            slices.push(AtomicHistogram::new(max, precision));
+            slices.push(Histogram::new(m, r, n));
             true_span += resolution;
         }
         slices.shrink_to_fit();
-        let next_tick = Instant::<Nanoseconds<AtomicU64>>::now();
+        let next_tick = AtomicInstant::now();
         next_tick.fetch_add(resolution, Ordering::Relaxed);
         Self {
             slices,
             current: AtomicUsize::new(0),
             next_tick,
             resolution,
-            summary: AtomicHistogram::new(max, precision),
+            summary: Histogram::new(m, r, n),
         }
     }
 
@@ -78,16 +64,11 @@ where
     }
 
     /// Increment a time-value pair by a specified count
-    pub fn increment(
-        &self,
-        time: Instant<Nanoseconds<u64>>,
-        value: Value,
-        count: <Count as Atomic>::Primitive,
-    ) {
+    pub fn increment(&self, time: Instant, value: u64, count: u32) {
         self.tick(time);
         if let Some(slice) = self.slices.get(self.current.load(Ordering::Relaxed)) {
-            slice.increment(value, count);
-            self.summary.increment(value, count);
+            let _ = slice.increment(value, count);
+            let _ = self.summary.increment(value, count);
         }
     }
 
@@ -104,16 +85,14 @@ where
     /// function. Users needing better consistency should ensure that other
     /// threads are not writing into the heatmap while this function is
     /// in-progress.
-    pub fn percentile(&self, percentile: f64) -> Result<Value, HeatmapError> {
-        self.tick(Instant::<Nanoseconds<u64>>::now());
-        self.summary
-            .percentile(percentile)
-            .map_err(HeatmapError::from)
+    pub fn percentile(&self, percentile: f64) -> Result<Bucket, Error> {
+        self.tick(Instant::now());
+        self.summary.percentile(percentile).map_err(Error::from)
     }
 
     // Internal function which handles reuse of older windows to store newer
     /// values.
-    fn tick(&self, time: Instant<Nanoseconds<u64>>) {
+    fn tick(&self, time: Instant) {
         loop {
             let next_tick = self.next_tick.load(Ordering::Relaxed);
             if time < next_tick {
@@ -126,74 +105,114 @@ where
                 }
                 let current = self.current.load(Ordering::Relaxed);
                 if let Some(slice) = self.slices.get(current) {
-                    self.summary.sub_assign(slice);
+                    let _ = self.summary.sub_assign(slice);
                     slice.clear();
                 }
             }
         }
     }
 
-    /// Performs a `Relaxed` load of the current `AtomicHeatmap` allocating and
-    /// returning a non-atomic `Heatmap`.
-    ///
-    /// Note: data may be inconsistent if there are concurrent writes happening
-    /// while the load operation is performed. Users needing better consistency
-    /// should ensure that other threads are not writing into the heatmap while
-    /// this operation is in-progress.
-    pub fn load(&self) -> Heatmap<Value, <Count as Atomic>::Primitive>
-    where
-        Value: Copy + std::ops::Sub<Output = Value>,
-        <Count as Atomic>::Primitive: Counter,
-    {
-        let mut result = Heatmap {
-            slices: Vec::with_capacity(self.slices.len()),
-            current: self.current.load(Ordering::Relaxed),
-            next_tick: self.next_tick.load(Ordering::Relaxed),
-            resolution: self.resolution,
-            summary: self.summary.load(),
-        };
-        for slice in &self.slices {
-            result.slices.push(slice.load());
+    /// Internal function to return a `Window` from the `Heatmap`.
+    fn get_slice(&self, index: usize) -> Option<Window> {
+        if let Some(histogram) = self.slices.get(index) {
+            let shift = if index > self.current.load(Ordering::Relaxed) {
+                self.resolution.mul_f64(
+                    (self.slices.len() + self.current.load(Ordering::Relaxed) - index) as f64,
+                )
+            } else {
+                self.resolution
+                    .mul_f64((self.current.load(Ordering::Relaxed) - index) as f64)
+            };
+            Some(Window {
+                start: self.next_tick.load(Ordering::Relaxed) - shift - self.resolution,
+                stop: self.next_tick.load(Ordering::Relaxed) - shift,
+                histogram,
+            })
+        } else {
+            None
         }
-        result.slices.shrink_to_fit();
-        result
+    }
+}
+
+impl Clone for Heatmap {
+    fn clone(&self) -> Self {
+        let slices = self.slices.clone();
+        let summary = self.summary.clone();
+        let resolution = self.resolution;
+        let current = AtomicUsize::new(self.current.load(Ordering::Relaxed));
+        let next_tick = AtomicInstant::new(self.next_tick.load(Ordering::Relaxed));
+
+        Heatmap {
+            slices,
+            current,
+            next_tick,
+            resolution,
+            summary,
+        }
+    }
+}
+
+pub struct Iter<'a> {
+    inner: &'a Heatmap,
+    index: usize,
+    visited: usize,
+}
+
+impl<'a> Iter<'a> {
+    fn new(inner: &'a Heatmap) -> Iter<'a> {
+        let index = if inner.current.load(Ordering::Relaxed) < (inner.slices.len() - 1) {
+            inner.current.load(Ordering::Relaxed) + 1
+        } else {
+            0
+        };
+        Iter {
+            inner,
+            index,
+            visited: 0,
+        }
+    }
+}
+
+impl<'a> Iterator for Iter<'a> {
+    type Item = Window<'a>;
+
+    fn next(&mut self) -> Option<Window<'a>> {
+        if self.visited >= self.inner.slices.len() {
+            None
+        } else {
+            let bucket = self.inner.get_slice(self.index);
+            self.index += 1;
+            if self.index >= self.inner.slices.len() {
+                self.index = 0;
+            }
+            self.visited += 1;
+            bucket
+        }
+    }
+}
+
+impl<'a> IntoIterator for &'a Heatmap {
+    type Item = Window<'a>;
+    type IntoIter = Iter<'a>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        Iter::new(self)
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use rustcommon_atomics::AtomicU64;
-
     use super::*;
 
     #[test]
     fn age_out() {
-        let mut heatmap = Heatmap::<u64, u64>::new(
-            1_000_000,
-            2,
-            Duration::<Nanoseconds<u64>>::from_secs(1),
-            Duration::<Nanoseconds<u64>>::from_millis(1),
-        );
-        assert_eq!(heatmap.percentile(0.0), Err(HeatmapError::Empty));
-        heatmap.increment(Instant::<Nanoseconds<u64>>::now(), 1, 1);
-        assert_eq!(heatmap.percentile(0.0), Ok(1));
+        let heatmap = Heatmap::new(0, 4, 20, Duration::from_secs(1), Duration::from_millis(1));
+        assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Err(Error::Empty));
+        heatmap.increment(Instant::now(), 1, 1);
+        assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Ok(1));
         std::thread::sleep(std::time::Duration::from_millis(100));
-        assert_eq!(heatmap.percentile(0.0), Ok(1));
+        assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Ok(1));
         std::thread::sleep(std::time::Duration::from_millis(2000));
-        assert_eq!(heatmap.percentile(0.0), Err(HeatmapError::Empty));
-
-        let heatmap = AtomicHeatmap::<u64, AtomicU64>::new(
-            1_000_000,
-            2,
-            Duration::<Nanoseconds<u64>>::from_secs(1),
-            Duration::<Nanoseconds<u64>>::from_millis(1),
-        );
-        assert_eq!(heatmap.percentile(0.0), Err(HeatmapError::Empty));
-        heatmap.increment(Instant::<Nanoseconds<u64>>::now(), 1, 1);
-        assert_eq!(heatmap.percentile(0.0), Ok(1));
-        std::thread::sleep(std::time::Duration::from_millis(100));
-        assert_eq!(heatmap.percentile(0.0), Ok(1));
-        std::thread::sleep(std::time::Duration::from_millis(2000));
-        assert_eq!(heatmap.percentile(0.0), Err(HeatmapError::Empty));
+        assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Err(Error::Empty));
     }
 }