--- a/heatmap/src/heatmap.rs
+++ b/heatmap/src/heatmap.rs
@@ -307,7 +307,7 @@ mod tests {
 
     #[test]
     fn age_out() {
-        let heatmap = Heatmap::new(0, 4, 20, Duration::from_secs(1), Duration::from_millis(1));
+        let heatmap = Heatmap::new(0, 4, 20, Duration::from_secs(1), Duration::from_millis(1)).unwrap();
         assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Err(Error::Empty));
         heatmap.increment(Instant::now(), 1, 1);
         assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Ok(1));