--- a/heatmap/src/heatmap.rs
+++ b/heatmap/src/heatmap.rs
@@ -115,7 +115,13 @@ impl Heatmap {
     ///
     /// - `resolution` - sets the resolution in the time domain. Counts from
     /// similar instants in time will be grouped together.
-    pub fn new(m: u32, r: u32, n: u32, span: Duration, resolution: Duration) -> Result<Self, Error> {
+    pub fn new(
+        m: u32,
+        r: u32,
+        n: u32,
+        span: Duration,
+        resolution: Duration,
+    ) -> Result<Self, Error> {
         let mut slices = Vec::new();
         let mut true_span = Duration::from_nanos(0);
         while true_span < span {
@@ -307,7 +313,8 @@ mod tests {
 
     #[test]
     fn age_out() {
-        let heatmap = Heatmap::new(0, 4, 20, Duration::from_secs(1), Duration::from_millis(1)).unwrap();
+        let heatmap =
+            Heatmap::new(0, 4, 20, Duration::from_secs(1), Duration::from_millis(1)).unwrap();
         assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Err(Error::Empty));
         heatmap.increment(Instant::now(), 1, 1);
         assert_eq!(heatmap.percentile(0.0).map(|v| v.high()), Ok(1));