--- a/heatmap/src/heatmaps/mod.rs
+++ b/heatmap/src/heatmaps/mod.rs
@@ -1,9 +1,0 @@
-// Copyright 2020 Twitter, Inc.
-// Licensed under the Apache License, Version 2.0
-// http://www.apache.org/licenses/LICENSE-2.0
-
-mod atomic;
-mod standard;
-
-pub use atomic::AtomicHeatmap;
-pub use standard::Heatmap;