--- a/histogram/.gitignore
+++ b/histogram/.gitignore
@@ -1,2 +1,0 @@
-/target
-Cargo.lock