--- a/histogram/CHANGELOG.md
+++ b/histogram/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Unreleased
-
-# 1.0.0 - 2020-09-01
-
-Initial release.