Replaces the histogram with a simple base-2 implementation which is opioniated about the values and counter widths. This simplifies usage by converging on a single implementation using atomic counters. Additionally, we get substantial performance benefits by switching to a base-2 implementation.

Similarly the heatmap crate is simplified to converge on a single implementation using atomics.