@brayniac 2022-09-29
Benchmark results:

```
Original Implementation:

AtomicHistogram/u64/AtomicU64/increment/min/precision/1
                        time:   [7.5811 ns 7.6297 ns 7.6806 ns]
                        thrpt:  [130.20 Melem/s 131.07 Melem/s 131.91 Melem/s]

AtomicHistogram/u64/AtomicU64/increment/max/precision/1
                        time:   [18.881 ns 18.996 ns 19.118 ns]
                        thrpt:  [52.308 Melem/s 52.643 Melem/s 52.962 Melem/s]

New Implementation:

histogram/increment/min time:   [6.1397 ns 6.2107 ns 6.2819 ns]
                        thrpt:  [159.19 Melem/s 161.01 Melem/s 162.87 Melem/s]

histogram/increment/max time:   [6.0612 ns 6.1219 ns 6.1839 ns]
                        thrpt:  [161.71 Melem/s 163.35 Melem/s 164.98 Melem/s]
```