@thinkingfish 2022-09-29
You can use your base of documentation and edit it as you wish https://github.com/thinkingfish/ccommon/blob/yao/2histogram-doc-md/docs/modules/2-histogram.md

I converted the original rst to markdown and fixed most of the styling issues.