@brayniac 2022-09-29
> You can use your base of documentation and edit it as you wish https://github.com/thinkingfish/ccommon/blob/yao/2histogram-doc-md/docs/modules/2-histogram.md
> 
> I converted the original rst to markdown using pandoc and fixed most of the styling issues.

Thanks! I'll circle back to add some better crate level docs before potentially publishing to crates.io 

For now, I think the rustdocs that I've added are sufficient to get a general sense of how to use this and I would like to get to testing the change in the context of Pelikan